move request from GenericBrowser to parent
